feat: Implement statistics management system with SQLite
- Added DatabaseManager class for handling SQLite database operations, including table creation, data insertion, and data retrieval.
- Created main script to manage the flow of data from dummy generation to statistics calculation and output.
- Developed oldStatistics.py for legacy support with extensive functionality for data validation, visitor counting, and logging.
- Introduced postprocessing and preprocessing modules for data filtering and validation.
- Implemented utility functions for generating dummy data for testing purposes.
- Established a clear structure for managing statistics, including filtering by stay time and visit counts.
